+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## v0.6.1 - ?
+## v0.7.0 - ?
 
+- Add git_ops::processors::get_template_value processor to convert template references to the rendered template.
+- Add automatic inmanta.plugins.Context injection to the attribute_processor decorator, passed as the first positional argument like the native plugin decorator does.
 
 ## v0.6.0 - 2026-06-17